offscreen-canvas/the-offscreen-canvas/offscreencanvas.transferrable.w.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="help" href="https://html.spec.whatwg.org/multipage/scripting.html#the-offscreen-2d-rendering-context">

<script id="myWorker" type="text/worker">

function isInvalidStateError(funcStr, offscreenCanvas)
{
    try {
        eval(funcStr);
    } catch (e) {
        if (e instanceof DOMException && e.name == "InvalidStateError")
            return true;
        return false;
    }
}

function test2()
{
    var offscreenCanvas = new OffscreenCanvas(10, 10);
    var ctx = offscreenCanvas.getContext('2d');
    return isInvalidStateError("self.postMessage(offscreenCanvas, [offscreenCanvas])", offscreenCanvas);
}

function test3()
{
    var offscreenCanvas = new OffscreenCanvas(10, 10);
    var ctx = offscreenCanvas.getContext('webgl');
    return isInvalidStateError("self.postMessage(offscreenCanvas, [offscreenCanvas])", offscreenCanvas);
}

function test4()
{
    var offscreenCanvas = new OffscreenCanvas(10, 10);
    self.postMessage(offscreenCanvas, [offscreenCanvas]);
    return isInvalidStateError("self.postMessage(offscreenCanvas, [offscreenCanvas])", offscreenCanvas);
}

function test5()
{
    var offscreenCanvas = new OffscreenCanvas(10, 10);
    self.postMessage(offscreenCanvas, [offscreenCanvas]);
    return isInvalidStateError("offscreenCanvas.getContext('2d')", offscreenCanvas);
}

function test6()
{
    var offscreenCanvas = new OffscreenCanvas(10, 10);
    self.postMessage(offscreenCanvas, [offscreenCanvas]);
    return isInvalidStateError("offscreenCanvas.getContext('webgl')", offscreenCanvas);
}

self.onmessage = function(e) {
    switch(e.data) {
        case 'test1':
            var offscreenCanvas = new OffscreenCanvas(10, 10);
            self.postMessage(offscreenCanvas, [offscreenCanvas]);
            break;
        case 'test2':
            self.postMessage(test2());
            break;
        case 'test3':
            self.postMessage(test3());
            break;
        case 'test4':
            self.postMessage(test4());
            break;
        case 'test5':
            self.postMessage(test5());
            break;
        case 'test6':
            self.postMessage(test6());
            break;
    }
};

</script>

<script>
function makeWorker(script)
{
    var blob = new Blob([script]);
    return new Worker(URL.createObjectURL(blob));
}

async_test(function(t) {
    var worker = makeWorker(document.getElementById("myWorker").textContent);
    worker.addEventListener('message', t.step_func_done(function(msg) {
        assert_equals(msg.data.width, 10);
        assert_equals(msg.data.height, 10);
    }));
    worker.postMessage('test1');
}, "test that OffscreenCanvas's size is correct after being transferred from a worker.");

async_test(function(t) {
    var worker = makeWorker(document.getElementById("myWorker").textContent);
    worker.addEventListener('message', t.step_func_done(function(msg) {
        assert_true(msg.data);
    }));
    worker.postMessage('test2');
}, "test that transfer an OffscreenCanvas that has a 2d context throws exception in a worker.");

async_test(function(t) {
    var worker = makeWorker(document.getElementById("myWorker").textContent);
    worker.addEventListener('message', t.step_func_done(function(msg) {
        assert_true(msg.data);
    }));
    worker.postMessage('test3');
}, "test that transfer an OffscreenCanvas that has a webgl context throws exception in a worker.");

async_test(function(t) {
    var worker = makeWorker(document.getElementById("myWorker").textContent);
    worker.addEventListener('message', t.step_func_done(function(msg) {
        if (msg.data == true || msg.data == false)
            assert_true(msg.data);
    }));
    worker.postMessage('test4');
}, "test that transfer an OffscreenCanvas twice throws exception in a worker.");

async_test(function(t) {
    var worker = makeWorker(document.getElementById("myWorker").textContent);
    worker.addEventListener('message', t.step_func_done(function(msg) {
        if (msg.data == true || msg.data == false)
            assert_true(msg.data);
    }));
    worker.postMessage('test5');
}, "test that call getContext('2d') on a detached OffscreenCanvas throws exception in a worker.");

async_test(function(t) {
    var worker = makeWorker(document.getElementById("myWorker").textContent);
    worker.addEventListener('message', t.step_func_done(function(msg) {
        if (msg.data == true || msg.data == false)
            assert_true(msg.data);
    }));
    worker.postMessage('test6');
}, "test that call getContext('webgl') on a detached OffscreenCanvas throws exception in a worker.");

</script>
